+++ app/Models/Brand.php
@@ -44,6 +44,7 @@ class Brand extends Model
         'status'=> BrandStatusEnums::class,
     ];
 
+
     public function users()
     {
         return $this->belongsToMany(User::class,'brand_user');