+++ database/factories/CouponFactory.php
@@ -25,6 +25,7 @@ class CouponFactory extends Factory
             'name' => 'coupon_name-'.Str::random(5),
             'brand_id'=> Brand::factory(),
 
+
             'max_redemptions'=> 1000,
             'max_discount_codes'=> 1000,
 
@@ -38,6 +39,7 @@ class CouponFactory extends Factory
 
             'discount_code_max_length'=> 10,
             'discount_code_type'=>CouponDiscountCodesTypeEnums::RANDOM_STRING->value,
+
             'discount_code_valid_for_max_hours_of'=>24,
             'trigger'=>'customer_created',
             'expires_at'=>now()->addDays(10)