+++ database/migrations/2022_05_08_003826_create_coupons_table.php
@@ -33,6 +33,7 @@ return new class extends Migration
 
 
             $table->integer('discount_code_max_length')->default('8');
+
             $table->tinyInteger('discount_code_type')->default(\App\Enums\CouponDiscountCodesTypeEnums::RANDOM_STRING->value);
             $table->integer('discount_codes_generated')->default(0);
             $table->integer('discount_codes_redeemed')->default(0);