+++ ChangeLog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## [0.0.2] - 2022-05-08
 ### Added
+- BootUpSeeder to build the main objects
 
 
 ## [0.0.1] - 2022-05-08