+++ ChangeLog.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
 ## [0.0.2] - 2022-05-08
 ### Added
 - BootUpSeeder to build the main objects
 - Feature Test for User Can get discount codes
 - User can get discount codes
+
 
 ## [0.0.1] - 2022-05-08
 ### Added